Refactor Style Form Dictionary
