+++ lib/assets/javascripts/builder/editor/style/style-form/style-form-default-model.js
@@ -31,7 +31,7 @@ module.exports = Backbone.Model.extend({
   _isTorqueCategory: function () {
     var HEATMAP_TYPE = 'heatmap';
     var fill = this._styleModel.get('fill');
-    var color = (fill && fill.color) || {};
+    var color = fill && fill.color || {};
 
     return this._styleModel.get('style') !== HEATMAP_TYPE
       ? !color.fixed
@@ -47,7 +47,7 @@ module.exports = Backbone.Model.extend({
     var styleType = this._styleModel.get('type');
     var isAutoStyleApplied = this._styleModel.has('autoStyle');
     var animationType = this._styleModel.get('style');
-    var formName = this._FORM_NAME ? (this._FORM_NAME + '-') : '';
+    var formName = this._FORM_NAME ? this._FORM_NAME + '-' : '';
 
     return _.reduce(this.attributes, function (attribute, value, key) {
       var formComponent = StyleFormComponents[formName + key];
