+++ lib/assets/javascripts/builder/editor/style/style-form/style-form-dictionary/aggregation-value.js
@@ -1,0 +1,15 @@
+var StyleFormDictionaryHelpers = require('builder/editor/style/style-form/style-form-helpers');
+
+module.exports = {
+  generate: function (params) {
+    return {
+      type: 'Operators',
+      title: _t('editor.style.components.aggregation-value.label'),
+      options: StyleFormDictionaryHelpers.getSchemaColumns(params.querySchemaModel),
+      dialogMode: 'float',
+      editorAttrs: {
+        help: _t('editor.style.components.aggregation-value.help')
+      }
+    };
+  }
+};
