+++ lib/assets/javascripts/builder/editor/style/style-form/style-form-dictionary/animated-duration.js
@@ -1,0 +1,16 @@
+module.exports = {
+  generate: function () {
+    return {
+      type: 'Number',
+      title: _t('editor.style.components.animated-duration.label'),
+      validators: ['required', {
+        type: 'interval',
+        min: 0,
+        max: 60
+      }],
+      editorAttrs: {
+        help: _t('editor.style.components.animated-duration.help')
+      }
+    };
+  }
+};
