+++ lib/assets/javascripts/builder/editor/style/style-form/style-form-dictionary/animated-overlap.js
@@ -1,0 +1,23 @@
+module.exports = {
+  generate: function (params) {
+    if (params.isTorqueCategory) {
+      return {
+        type: 'Hidden'
+      };
+    }
+
+    return {
+      type: 'Radio',
+      title: _t('editor.style.components.animated-overlap.label'),
+      options: [
+        {
+          val: 'false',
+          label: _t('editor.style.components.animated-overlap.options.false')
+        }, {
+          val: 'true',
+          label: _t('editor.style.components.animated-overlap.options.true')
+        }
+      ]
+    };
+  }
+};
