+++ lib/assets/javascripts/builder/editor/style/style-form/style-form-dictionary/animated-steps.js
@@ -1,0 +1,17 @@
+module.exports = {
+  generate: function () {
+    return {
+      type: 'Number',
+      title: _t('editor.style.components.animated-steps.label'),
+      validators: ['required', {
+        type: 'interval',
+        min: 1,
+        max: 1024,
+        step: 4
+      }],
+      editorAttrs: {
+        help: _t('editor.style.components.animated-steps.help')
+      }
+    };
+  }
+};
