+++ lib/assets/javascripts/builder/editor/style/style-form/style-form-dictionary/hidden.js
@@ -1,0 +1,7 @@
+module.exports = {
+  generate: function () {
+    return {
+      type: 'Hidden'
+    };
+  }
+};
