+++ lib/assets/javascripts/builder/editor/style/style-form/style-form-dictionary/labels-fill.js
@@ -1,0 +1,21 @@
+module.exports = {
+  generate: function () {
+    return {
+      type: 'Fill',
+      title: _t('editor.style.components.labels-fill'),
+      options: [],
+      dialogMode: 'float',
+      editorAttrs: {
+        size: {
+          min: 6,
+          max: 24,
+          hidePanes: ['value']
+        },
+        color: {
+          hidePanes: ['value']
+        }
+      },
+      validators: ['required']
+    };
+  }
+};
