+++ lib/assets/javascripts/builder/editor/style/style-form/style-form-dictionary/labels-font.js
@@ -1,0 +1,23 @@
+var FONTS_LIST = [
+  'DejaVu Sans Book',
+  'Unifont Medium',
+  'Open Sans Regular',
+  'Lato Regular',
+  'Lato Bold',
+  'Lato Bold Italic',
+  'Graduate Regular',
+  'Gravitas One Regular',
+  'Old Standard TT Regular',
+  'Old Standard TT Italic',
+  'Old Standard TT Bold'
+];
+
+module.exports = {
+  generate: function () {
+    return {
+      type: 'Select',
+      dialogMode: 'float',
+      options: FONTS_LIST
+    };
+  }
+};
