+++ lib/assets/javascripts/builder/editor/style/style-form/style-form-dictionary/labels-halo.js
@@ -1,0 +1,19 @@
+module.exports = {
+  generate: function () {
+    return {
+      type: 'Fill',
+      title: _t('editor.style.components.labels-halo'),
+      options: [],
+      dialogMode: 'float',
+      editorAttrs: {
+        size: {
+          hidePanes: ['value']
+        },
+        color: {
+          hidePanes: ['value']
+        }
+      },
+      validators: ['required']
+    };
+  }
+};
