+++ lib/assets/javascripts/builder/editor/style/style-form/style-form-dictionary/labels-offset.js
@@ -1,0 +1,13 @@
+module.exports = {
+  generate: function () {
+    return {
+      type: 'Number',
+      title: _t('editor.style.components.labels-offset'),
+      validators: ['required', {
+        type: 'interval',
+        min: -15,
+        max: 15
+      }]
+    };
+  }
+};
