+++ lib/assets/javascripts/builder/editor/style/style-form/style-form-dictionary/labels-overlap.js
@@ -1,0 +1,17 @@
+module.exports = {
+  generate: function () {
+    return {
+      type: 'Radio',
+      title: _t('editor.style.components.labels-overlap.label'),
+      options: [
+        {
+          val: 'true',
+          label: _t('editor.style.components.labels-overlap.options.true')
+        }, {
+          val: 'false',
+          label: _t('editor.style.components.labels-overlap.options.false')
+        }
+      ]
+    };
+  }
+};
