+++ lib/assets/javascripts/builder/editor/style/style-form/style-form-dictionary/stroke-size.js
@@ -1,0 +1,5 @@
+module.exports = {
+  generate: function () {
+    return {}; // TODO
+  }
+};
